+++ src/lib.rs
@@ -3,6 +3,6 @@ extern crate ndarray;
 extern crate fftw3_sys as ffi;
 extern crate num_complex;
 
-pub mod types;
-pub mod one_dim;
 pub mod prelude;
+pub mod traits;
+pub mod impl_vec;
