+++ src/prelude.rs
@@ -1,2 +1,2 @@
-pub use one_dim::RealArray as RealArray1D;
-pub use one_dim::ComplexArray as ComplexArray1D;
+pub use traits::RealArray;
+pub use traits::ComplexArray;
